Force shippable to use a Python build environment
As per the docs, a language must be chosen or Shippable will choose one
for you which may not be what you need and can cause your build to fail.
Example, this project is not a Ruby project but Shippable will consider
it one unless I specify Python. Without me specifying Python, Shippable
will try to find a Gemfile in the root dir and fail for not finding one.